Add combined

+++ java/shared/src/main/java/cloud/yogurt/shared/header/HeaderIntegerValue.java
@@ -1,16 +1,16 @@
 package cloud.yogurt.shared.header;
 
 public class HeaderIntegerValue extends HeaderValue {
-    private int value;
+    private long value;
 
-    public HeaderIntegerValue(int value) {
+    public HeaderIntegerValue(long value) {
         this.value = value;
     }
     @Override
     public String toString() {
         return "" + this.value;
     }
-    public int getValue() {
+    public long getValue() {
         return value;
     }
 }
